Move static funcs to utils file p1
